terminal client: specify the app clock granularity
the app clock granularity is different on different systems (100ms for
the readline interface, 50 ms for the plain command line, 0.5ms for the
daemon mode). this patch sets the granularity to 20ms (50 ticks per second)
for all different systems.

as a side effect it reduces the effect of a memory leak.

Signed-off-by: Tim Blechmann <tim@klingt.org>

git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/branches/3.4@10096 a380766d-ff14-0410-b294-a243070f3f08