sclang: force size of PyrSlot to 16 byte and fix PyrMethodRaw size
we manually increase the size of PyrSlot to 16 bytes. while this wastes
some space, it avoids that the compiler will automatically pad the
struct for performance reasons. then we also change PyrMethodRaw to match
the size of PyrSlot[2], so that it can be safely store.
the best way would be to use a union to extract PyrMethodRaw from the
PyrSlots of PyrObjects. However this would require some more code
restructuring

Signed-off-by: Tim Blechmann <tim@klingt.org>