Improve performance of sclang-insert-method-args
The SC code to obtain the list of classes responding to the method
selector was spectacularly inefficient. It iterated over all classes
and used findRespondingMethodFor, which looks upward through superclasses
if a method is inherited. This means that superclasses could be touched
an unknown number of times.

The revision walks recursively through the class tree. If a class responds
to the method selector, then we can assume all of its subclasses also
respond to that selector. (We only need the list of class names, not
the specific responding method.)

This reduces the typical search time from 2.0-2.5 seconds,
to 0.05-0.08 seconds.