sclang: increase udp receive buffer size
the osc parser of sclang assumes that the full message is placed in the
buffer. if this is not the case, it crashes. this happens e.g. when
replying to g_queryTree when may synths are running.

this patch is merely a workaround, as it just increases the buffer size
and moves the buffer to the SC_UdpInPort and SC_UdpCustomInPort classes

Signed-off-by: Tim Blechmann <tim@klingt.org>