sclang: bind Ctrl-T to CmdPeriod
Previously, Ctrl-B was bound, but on Windows, the same ASCII code
appears to be produced by the left arrow key, which cripples the
command line navigation.

Instead, bind Ctrl-T. There's several reasons to support that:
1. it makes sense lexically (as in "Terminate");
2. it does not appear to be bound to a signal by any standard;
3. it is not expected to be bound to another important terminal
feature, as for example backward navigation - it is described as
intended for "stopping an auxiliary device".
4. 'T' is usually far enough on keyboards from 'X', that we use
for library recompilation