Set Server.program only in *Platform's startup. Don't change cwd of scsynth.
There was a mix of trying to set Server.program in Server *initClass (with
Platform.switch, which does not exist) and in the Platform-specific classes.
We also remove the "cd" before starting scsynth, so that it will inherit the
CWD of the lang.
(cherry picked from commit faf77146a8634acdd1286b82df73dd0a5f52a2e8)